backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==24.10.0
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-debug-toolbar==4.4.6
dj-database-url==2.3.0
django-cors-headers==4.7.0
django-celery-beat==2.7.0
django-extensions==3.2.3
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
django-redis==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
phonenumbers==8.13.52
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
hiredis==3.1.0
celery==5.4.0
freezegun==1.5.1
gunicorn==23.0.0
django-storages[s3]==1.14.5
python-decouple==3.8
boto3==1.36.25
sentry-sdk[django]==2.20.0
